+++ demos/modules/demo_chart.mjs
@@ -3,21 +3,21 @@
  * AUTH: Brent Ely (https://github.com/gitbrent/)
  * DESC: Common test/demo slides for all library features
  * DEPS: Used by various demos (./demos/browser, ./demos/node, etc.)
- * VER.: 3.10.0
- * BLD.: 20220224
+ * VER.: 3.11.0
+ * BLD.: 20220515
  */
 
 import {
-	IMAGE_PATHS,
 	BASE_TABLE_OPTS,
 	BASE_TEXT_OPTS_L,
 	BASE_TEXT_OPTS_R,
-	COLOR_RED,
 	COLOR_AMB,
-	COLOR_GRN,
-	COLOR_UNK,
-	TESTMODE,
 	COLOR_BLU,
+	COLOR_GRN,
+	COLOR_RED,
+	COLOR_UNK,
+	IMAGE_PATHS,
+	TESTMODE,
 } from "./enums.mjs";
 
 const LETTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ".split("");
@@ -127,8 +127,9 @@ function genSlide01(pptx) {
 		altText: "this is the alt text content",
 
 		barDir: "bar",
-		border: { pt: "3", color: "00EE00" },
-		fill: "F1F1F1",
+		//plotArea: { border: { pt: "3", color: "00EE00" }, fill: { color: "F1F1F1" } },
+		border: { pt: "3", color: "00EE00" }, // @deprecated - legacy text only (dont use this syntax - use `plotArea`)
+		fill: "F1F1F1", // @deprecated - legacy text only (dont use this syntax - use `plotArea`)
 
 		catAxisLabelColor: "CC0000",
 		catAxisLabelFontFace: "Helvetica Neue",
@@ -137,7 +138,7 @@ function genSlide01(pptx) {
 		catAxisMajorTickMark: "in",
 		catAxisMinorTickMark: "cross",
 
-		// valAxisCrossesAt: 10,
+		//valAxisCrossesAt: 100,
 		valAxisMajorTickMark: "cross",
 		valAxisMinorTickMark: "out",
 
@@ -182,7 +183,6 @@ function genSlide01(pptx) {
 	slide.addChart(pptx.charts.BAR, arrDataRegions, optsChartBar2);
 
 	// BTM-LEFT: H/bar - TITLE and LEGEND
-	slide.addText(".", { x: 0.5, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartBar3 = {
 		x: 0.5,
 		y: 3.8,
@@ -190,8 +190,12 @@ function genSlide01(pptx) {
 		h: 3.5,
 		barDir: "bar",
 
-		border: { pt: "3", color: "CF0909" },
-		fill: "F1C1C1",
+		//chartArea: { fill: { color: "F1F1F1" } },
+		//chartArea: { fill: { color: pptx.colors.BACKGROUND2 } },
+		chartArea: { fill: { color: pptx.colors.BACKGROUND2 }, border: { color: pptx.colors.ACCENT3, pt: 3 } },
+		//chartArea: { fill: { color: "F1F1F1", transparency: 75 } },
+		plotArea: { border: { pt: "3", color: "CF0909" }, fill: { color: "F1C1C1" } },
+		//plotArea: { border: { pt: "3", color: "CF0909" }, fill: { color: "F1C1C1", transparency: 10 } },
 
 		catAxisLabelColor: "CC0000",
 		catAxisLabelFontFace: "Helvetica Neue",
@@ -208,12 +212,13 @@ function genSlide01(pptx) {
 	slide.addChart(pptx.charts.BAR, arrDataHighVals, optsChartBar3);
 
 	// BTM-RIGHT: V/col - TITLE and LEGEND
-	slide.addText(".", { x: 7.0, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartBar4 = {
 		x: 7.0,
 		y: 3.8,
 		w: 6.0,
 		h: 3.5,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		barDir: "col",
 		barGapWidthPct: 25,
 		chartColors: ["0088CC", "99FFCC"],
@@ -280,7 +285,7 @@ function genSlide02(pptx) {
 		w: 6.0,
 		h: 3.0,
 		barDir: "bar",
-		fill: "F1F1F1",
+		plotArea: { fill: { color: "F1F1F1" } },
 
 		catAxisLabelColor: "CC0000",
 		catAxisLabelFontFace: "Helvetica Neue",
@@ -309,7 +314,7 @@ function genSlide02(pptx) {
 		w: 6.0,
 		h: 3.0,
 		barDir: "col",
-		fill: "E1F1FF",
+		plotArea: { fill: { color: "E1F1FF" } },
 
 		dataBorder: { pt: "1", color: "F1F1F1" },
 		dataLabelColor: "696969",
@@ -332,17 +337,17 @@ function genSlide02(pptx) {
 	slide.addChart(pptx.charts.BAR, arrDataRegions, optsChartBar2);
 
 	// BTM-LEFT: H/bar - TITLE and LEGEND
-	slide.addText(".", { x: 0.5, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartBar3 = {
 		x: 0.5,
 		y: 3.8,
 		w: 6.0,
 		h: 3.5,
+
+		chartArea: { fill: { color: "F1F1F1" } },
+		plotArea: { border: { pt: "3", color: "CF0909" }, fill: { color: "F1C1C1" } },
+
 		barDir: "bar",
 		barOverlapPct: -50,
-
-		border: { pt: "3", color: "CF0909" },
-		fill: "F1C1C1",
 
 		catAxisLabelColor: "CC0000",
 		catAxisLabelFontFace: "Helvetica Neue",
@@ -353,11 +358,11 @@ function genSlide02(pptx) {
 		catAxisTitleFontSize: 14,
 		showCatAxisTitle: true,
 
-		valAxisOrientation: "maxMin",
+		catGridLine: { color: "cc6699", style: "dash", size: 1 },
 		valGridLine: { style: "none" },
+		valAxisOrientation: "maxMin",
 		valAxisHidden: true,
 		valAxisDisplayUnitLabel: true,
-		catGridLine: { color: "cc6699", style: "dash", size: 1 },
 
 		titleColor: "33CF22",
 		titleFontFace: "Helvetica Neue",
@@ -369,12 +374,13 @@ function genSlide02(pptx) {
 	slide.addChart(pptx.charts.BAR, arrDataHighVals, optsChartBar3);
 
 	// BTM-RIGHT: V/col - TITLE and LEGEND
-	slide.addText(".", { x: 7.0, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartBar4 = {
 		x: 7.0,
 		y: 3.8,
 		w: 6.0,
 		h: 3.5,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		barDir: "col",
 		barGapWidthPct: 25,
 		chartColors: ["0088CC", "99FFCC"],
@@ -512,12 +518,13 @@ function genSlide03(pptx) {
 	slide.addChart(pptx.charts.BAR, arrDataRegions, optsChartBar3);
 
 	// BTM-RIGHT: V/col - TITLE and LEGEND
-	slide.addText(".", { x: 7.0, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartBar4 = {
 		x: 7.0,
 		y: 3.8,
 		w: 6.0,
 		h: 3.5,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		barDir: "col",
 		barGrouping: "percentStacked",
 
@@ -612,6 +619,10 @@ function genSlide05(pptx) {
 			y: 0.6,
 			w: "45%",
 			h: 3,
+			chartArea: { fill: { color: "404040" } },
+			catAxisLabelColor: "F1F1F1",
+			valAxisLabelColor: "F1F1F1",
+
 			barDir: "bar",
 			chartColors: ["0077BF", "4E9D2D", "ECAA00", "5FC4E3", "DE4216", "154384"],
 			catLabelFormatCode: "yyyy-mm",
@@ -642,10 +653,16 @@ function genSlide05(pptx) {
 			y: 0.6,
 			w: "45%",
 			h: 3,
+			chartArea: { fill: { color: "404040" } },
+			catAxisLabelColor: "F1F1F1",
+			valAxisLabelColor: "F1F1F1",
+			valAxisLineColor: "7F7F7F",
+			valGridLine: { color: "7F7F7F" },
+			dataLabelColor: "B7B7B7",
+
 			valAxisMaxVal: 1,
 			barDir: "bar",
 			catAxisLineShow: false,
-			valAxisLineShow: false,
 			showValue: true,
 			catLabelFormatCode: "mmm-yy",
 			dataLabelPosition: "outEnd",
@@ -672,6 +689,13 @@ function genSlide05(pptx) {
 			y: 4.0,
 			w: "45%",
 			h: 3,
+			chartArea: { fill: { color: "404040" } },
+			catAxisLabelColor: "F1F1F1",
+			valAxisLabelColor: "F1F1F1",
+			valAxisLineColor: "7F7F7F",
+			valGridLine: { color: "7F7F7F" },
+			dataLabelColor: "B7B7B7",
+
 			barDir: "col", // `col`(vert) | `bar`(horiz)
 			showValue: true,
 			dataLabelPosition: "outEnd",
@@ -682,6 +706,7 @@ function genSlide05(pptx) {
 			barGapWidthPct: 50,
 			showLegend: true,
 			legendPos: "r",
+			legendColor: "F1F1F1",
 		}
 	);
 
@@ -700,6 +725,13 @@ function genSlide05(pptx) {
 			y: 4,
 			w: "45%",
 			h: 3,
+			chartArea: { fill: { color: "404040" } },
+			catAxisLabelColor: "F1F1F1",
+			valAxisLabelColor: "F1F1F1",
+			valAxisLineColor: "7F7F7F",
+			valGridLine: { color: "7F7F7F" },
+			dataLabelColor: "B7B7B7",
+
 			barDir: "bar",
 			showValue: true,
 			dataLabelPosition: "outEnd",
@@ -750,8 +782,9 @@ function genSlide06(pptx) {
 		y: 0.6,
 		w: 6.0,
 		h: 3.0,
+		chartArea: { fill: { color: "F1F1F1", transparency: 50 } },
+
 		barDir: "bar",
-		fill: "F1F1F1",
 
 		catAxisLabelColor: pptx.colors.ACCENT2,
 		catAxisLabelFontFace: "Arial",
@@ -771,6 +804,8 @@ function genSlide06(pptx) {
 		y: 0.6,
 		w: 6.0,
 		h: 3.0,
+		chartArea: { fill: { color: "F1F1F1", transparency: 50 } },
+
 		barDir: "col",
 		bar3DShape: "cylinder",
 		catAxisLabelColor: "0000CC",
@@ -788,12 +823,13 @@ function genSlide06(pptx) {
 	slide.addChart(pptx.charts.BAR3D, arrDataRegions, optsChartBar2);
 
 	// BTM-LEFT: H/bar - TITLE and LEGEND
-	slide.addText(".", { x: 0.5, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartBar3 = {
 		x: 0.5,
 		y: 3.8,
 		w: 6.0,
 		h: 3.5,
+		chartArea: { fill: { color: "F1F1F1", transparency: 50 } },
+
 		barDir: "col",
 		bar3DShape: "pyramid",
 		barGrouping: "stacked",
@@ -811,12 +847,13 @@ function genSlide06(pptx) {
 	slide.addChart(pptx.charts.BAR3D, arrDataHighVals, optsChartBar3);
 
 	// BTM-RIGHT: V/col - TITLE and LEGEND
-	slide.addText(".", { x: 7.0, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartBar4 = {
 		x: 7.0,
 		y: 3.8,
 		w: 6.0,
 		h: 3.5,
+		chartArea: { fill: { color: "F1F1F1", transparency: 50 } },
+
 		barDir: "col",
 		bar3DShape: "coneToMax",
 		chartColors: ["0088CC", "99FFCC"],
@@ -867,6 +904,8 @@ function genSlide07(pptx) {
 			y: 0.5,
 			w: "90%",
 			h: "90%",
+			chartArea: { fill: { color: "F1F1F1", transparency: 50 } },
+
 			valAxisMaxVal: 1,
 			barDir: "bar",
 			axisLabelFormatCode: "#%",
@@ -895,12 +934,13 @@ function genSlide08(pptx) {
 		BASE_TABLE_OPTS
 	);
 
-	slide.addText("..", { x: 0.5, y: 0.6, w: 6.0, h: 3.0, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	let optsChartLine1 = {
 		x: 0.5,
 		y: 0.6,
 		w: 6.0,
 		h: 3.0,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		chartColors: [COLOR_RED, COLOR_AMB, COLOR_GRN, COLOR_UNK],
 		lineSize: 8,
 		lineSmooth: true,
@@ -1010,10 +1050,8 @@ function genSlide10(pptx) {
 		y: 0.6,
 		w: "95%",
 		h: "85%",
-		fill: "F2F9FC",
-
+		plotArea: { fill: { color: "F2F9FC" } },
 		valAxisMaxVal: MAXVAL,
-
 		showLegend: true,
 		legendPos: "r",
 	};
@@ -1052,8 +1090,9 @@ function genSlide11(pptx) {
 		y: 0.6,
 		w: "45%",
 		h: 3,
+		plotArea: { fill: { color: "D1E1F1" } },
+
 		catAxisLabelRotate: 45,
-		fill: "D1E1F1",
 		chartColors: ["0088CC"],
 		chartColorsOpacity: 25,
 		dataBorder: { pt: 2, color: "FFFFFF" },
@@ -1067,12 +1106,13 @@ function genSlide11(pptx) {
 		y: 0.6,
 		w: "45%",
 		h: 3,
+		plotArea: { fill: { color: "D1E1F1" } },
+
 		chartColors: ["0088CC", "99FFCC"],
 		chartColorsOpacity: 25,
 		valAxisLabelRotate: 5,
 		dataBorder: { pt: 2, color: "FFFFFF" },
 		showValue: false,
-		fill: "D1E1F1",
 		barGrouping: "stacked",
 	};
 	slide.addChart(pptx.charts.AREA, arrDataTimeline2ser, optsChartLine2);
@@ -1111,15 +1151,17 @@ function genSlide12(pptx) {
 		});
 
 	// TOP-LEFT
-	slide.addText(".", { x: 0.5, y: 0.5, w: 4.2, h: 3.2, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	slide.addChart(pptx.charts.PIE, dataChartPieStat, {
 		x: 0.5,
 		y: 0.5,
 		w: 4.2,
 		h: 3.2,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		legendPos: "left",
 		legendFontFace: "Courier New",
 		showLegend: true,
+
 		showLeaderLines: true,
 		showPercent: false,
 		showValue: true,
@@ -1131,25 +1173,47 @@ function genSlide12(pptx) {
 	});
 
 	// TOP-MIDDLE
-	slide.addText(".", { x: 5.6, y: 0.5, w: 3.2, h: 3.2, fill: { color: "F1F1F1" }, color: "F1F1F1" });
-	slide.addChart(pptx.charts.PIE, dataChartPieStat, { x: 5.6, y: 0.5, w: 3.2, h: 3.2, showLegend: true, legendPos: "t" });
+	slide.addChart(pptx.charts.PIE, dataChartPieStat, {
+		x: 5.6,
+		y: 0.5,
+		w: 3.2,
+		h: 3.2,
+		chartArea: { fill: { color: "F1F1F1" } },
+		showLegend: true,
+		legendPos: "t",
+	});
 
 	// BTM-LEFT
-	slide.addText(".", { x: 0.5, y: 4.0, w: 4.2, h: 3.2, fill: { color: "F1F1F1" }, color: "F1F1F1" });
-	slide.addChart(pptx.charts.PIE, dataChartPieLocs, { x: 0.5, y: 4.0, w: 4.2, h: 3.2, showLegend: true, legendPos: "r" });
+	slide.addChart(pptx.charts.PIE, dataChartPieLocs, {
+		x: 0.5,
+		y: 4.0,
+		w: 4.2,
+		h: 3.2,
+		chartArea: { fill: { color: "F1F1F1" } },
+		showLegend: true,
+		legendPos: "r",
+	});
 
 	// BTM-MIDDLE
-	slide.addText(".", { x: 5.6, y: 4.0, w: 3.2, h: 3.2, fill: { color: "F1F1F1" }, color: "F1F1F1" });
-	slide.addChart(pptx.charts.PIE, dataChartPieLocs, { x: 5.6, y: 4.0, w: 3.2, h: 3.2, showLegend: true, legendPos: "b" });
+	slide.addChart(pptx.charts.PIE, dataChartPieLocs, {
+		x: 5.6,
+		y: 4.0,
+		w: 3.2,
+		h: 3.2,
+		chartArea: { fill: { color: "F1F1F1" } },
+		showLegend: true,
+		legendPos: "b",
+	});
 
 	// BOTH: TOP-RIGHT
 	// DEMO: `legendFontSize`, `titleAlign`, `titlePos`
-	slide.addText(".", { x: 9.8, y: 0.5, w: 3.2, h: 3.2, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	slide.addChart(pptx.charts.PIE, dataChartPieLocs, {
 		x: 9.8,
 		y: 0.5,
 		w: 3.2,
 		h: 3.2,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		dataBorder: { pt: "1", color: "F1F1F1" },
 		showLegend: true,
 		legendPos: "t",
@@ -1162,12 +1226,13 @@ function genSlide12(pptx) {
 	});
 
 	// BOTH: BTM-RIGHT
-	slide.addText(".", { x: 9.8, y: 4.0, w: 3.2, h: 3.2, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	slide.addChart(pptx.charts.PIE, dataChartPieLocs, {
 		x: 9.8,
 		y: 4.0,
 		w: 3.2,
 		h: 3.2,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		dataBorder: { pt: "1", color: "F1F1F1" },
 		showLegend: true,
 		legendPos: "b",
@@ -1188,6 +1253,8 @@ function genSlide13(pptx) {
 		y: 1.0,
 		w: 6.0,
 		h: 6.0,
+		chartArea: { fill: { color: "F1F1F1" } },
+
 		chartColors: ["FC0000", "FFCC00", "009900", "0088CC", "696969", "6600CC"],
 		dataBorder: { pt: "2", color: "F1F1F1" },
 		dataLabelColor: "FFFFFF",
@@ -1208,7 +1275,6 @@ function genSlide13(pptx) {
 		titleFontFace: "Helvetica Neue",
 		titleFontSize: 24,
 	};
-	slide.addText(".", { x: 0.5, y: 1.0, w: 6.0, h: 6.0, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	slide.addChart(pptx.charts.DOUGHNUT, dataChartPieStat, optsChartPie1);
 
 	let optsChartPie2 = {
@@ -1282,7 +1348,8 @@ function genSlide14(pptx) {
 		y: 0.6,
 		w: "45%",
 		h: 3,
-		fill: "f1f1f1",
+		plotArea: { fill: { color: "F1F1F1" } },
+
 		showLegend: true,
 		legendPos: "b",
 
@@ -1302,7 +1369,8 @@ function genSlide14(pptx) {
 		y: 4.0,
 		w: "45%",
 		h: 3,
-		fill: "f2f9fc",
+		plotArea: { fill: { color: "F2F9FC" } },
+
 		//catAxisOrientation: 'maxMin',
 		//valAxisOrientation: 'maxMin',
 		showLegend: true,
@@ -1336,14 +1404,14 @@ function genSlide15(pptx) {
 
 	let arrDataBubble1 = [
 		{ name: "X-Axis", values: [0.3, 0.6, 0.9, 1.2, 1.5, 1.7] },
-		{ name: "Y-Value 1", values: [1.3, 9, 7.5, 2.5, 7.5, 5], sizes: [1, 4, 2, 3, 7, 4] },
-		{ name: "Y-Value 2", values: [5, 3, 2, 7, 2, 10], sizes: [9, 7, 9, 2, 4, 8] },
+		{ name: "Y-Value 1", values: [1.3, 9, 7.5, 2.5, 7.5, 3], sizes: [1, 4, 2, 3, 7, 4] },
+		{ name: "Y-Value 2", values: [5.0, 3, 2.0, 7.0, 2.0, 9], sizes: [9, 7, 9, 2, 4, 8] },
 	];
 	let arrDataBubble2 = [
 		{ name: "X-Axis", values: [1, 2, 3, 4, 5, 6] },
 		{ name: "Airplane", values: [33, 20, 51, 65, 71, 75], sizes: [10, 10, 12, 12, 15, 20] },
 		{ name: "Train", values: [99, 88, 77, 89, 99, 99], sizes: [20, 20, 22, 22, 25, 30] },
-		{ name: "Bus", values: [21, 22, 25, 49, 59, 69], sizes: [11, 11, 13, 13, 16, 21] },
+		{ name: "Bus", values: [21, 25, 32, 49, 59, 69], sizes: [11, 11, 13, 13, 16, 21] },
 	];
 
 	// TOP-LEFT
@@ -1352,11 +1420,11 @@ function genSlide15(pptx) {
 		y: 0.6,
 		w: "45%",
 		h: 3,
+		chartArea: { fill: { color: "F1F1F1" } },
 		chartColors: ["4477CC", "ED7D31"],
 		chartColorsOpacity: 40,
 		dataBorder: { pt: 1, color: "FFFFFF" },
 	};
-	slide.addText(".", { x: 0.5, y: 0.6, w: 6.0, h: 3.0, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 	slide.addChart(pptx.charts.BUBBLE, arrDataBubble1, optsChartBubble1);
 
 	// TOP-RIGHT
@@ -1365,7 +1433,8 @@ function genSlide15(pptx) {
 		y: 0.6,
 		w: "45%",
 		h: 3,
-		fill: "f1f1f1",
+		plotArea: { fill: { color: "F1F1F1" } },
+
 		showLegend: true,
 		legendPos: "b",
 
@@ -1385,21 +1454,31 @@ function genSlide15(pptx) {
 		y: 4.0,
 		w: "45%",
 		h: 3,
-		fill: "f2f9fc",
+		chartArea: { fill: { color: "404040" } },
+		plotArea: { fill: { color: "202020" } },
+
+		catAxisLabelColor: "F1F1F1",
+		catAxisLabelFontSize: 10,
 		catAxisOrientation: "maxMin",
-		valAxisOrientation: "maxMin",
 		showCatAxisTitle: false,
-		showValAxisTitle: false,
+		//
+		valAxisLabelColor: "F1F1F1",
+		valAxisLabelFontSize: 10,
 		valAxisMinVal: 0,
-		dataBorder: { pt: 2, color: "FFFFFF" },
-		dataLabelColor: "FFFFFF",
+		valAxisOrientation: "maxMin",
+		showValAxisTitle: false,
+		//
+		dataBorder: { pt: 2, color: "e1e1e1" },
+		dataLabelFontFace: "Arial",
+		dataLabelFontSize: 10,
+		dataLabelColor: "e1e1e1",
 		showValue: true,
 	};
 	slide.addChart(pptx.charts.BUBBLE, arrDataBubble1, optsChartBubble3);
 
 	// BOTTOM-RIGHT
 	let optsChartBubble4 = { x: 7.0, y: 4.0, w: "45%", h: 3, lineSize: 0 };
-	slide.addChart(pptx.charts.BUBBLE, arrDataBubble2, optsChartBubble4);
+	slide.addChart(pptx.charts.BUBBLE3D, arrDataBubble2, optsChartBubble4);
 }
 
 // SLIDE 16: Radar Chart
@@ -1436,12 +1515,13 @@ function genSlide16(pptx) {
 	// TOP-ROW
 	{
 		// TOP-L: `{ radar:'normal' }`
-		slide.addText("....", { x: 0.5, y: 0.6, w: 4.0, h: 3.0, fill: { color: "F9F9F9" }, color: "F9F9F9" });
 		let optsChartRadar1 = {
 			x: 0.5,
 			y: 0.6,
 			w: 4.0,
 			h: 3.0,
+			chartArea: { fill: { color: "F9F9F9" } },
+			//
 			radarStyle: "standard",
 			//
 			showTitle: true,
@@ -1455,12 +1535,13 @@ function genSlide16(pptx) {
 		slide.addChart(pptx.charts.RADAR, arrDataRegions, optsChartRadar1);
 
 		// TOP-C: `{ radar:'marker' }` Cat Axis options
-		slide.addText("....", { x: 4.65, y: 0.6, w: 4.0, h: 3.0, fill: { color: "F9F9F9" }, color: "F9F9F9" });
 		let optsChartRadar2 = {
 			x: 4.65,
 			y: 0.6,
 			w: 4.0,
 			h: 3.0,
+			chartArea: { fill: { color: "F9F9F9" } },
+			//
 			radarStyle: "marker",
 			//
 			showTitle: true,
@@ -1472,12 +1553,13 @@ function genSlide16(pptx) {
 		slide.addChart(pptx.charts.RADAR, arrDataRegions, optsChartRadar2);
 
 		// TOP-R: `{ radar:'marker' }` Cat Axis options
-		slide.addText("....", { x: 8.8, y: 0.6, w: 4.0, h: 3.0, fill: { color: "F9F9F9" }, color: "F9F9F9" });
 		let optsChartRadar3 = {
 			x: 8.8,
 			y: 0.6,
 			w: 4.0,
 			h: 3.0,
+			chartArea: { fill: { color: "F9F9F9" } },
+			//
 			radarStyle: "filled",
 			//
 			showTitle: true,
@@ -1492,12 +1574,13 @@ function genSlide16(pptx) {
 	// BTM-ROW
 	{
 		// BTM-L: marker/line options
-		slide.addText(".", { x: 0.5, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 		let optsChartRadar10 = {
 			x: 0.5,
 			y: 3.8,
 			w: 6.0,
 			h: 3.5,
+			chartArea: { fill: { color: "F1F1F1" } },
+			//
 			radarStyle: "marker",
 			catAxisLabelColor: COLOR_BLU,
 			catAxisLabelFontFace: "Courier",
@@ -1524,12 +1607,13 @@ function genSlide16(pptx) {
 		slide.addChart(pptx.charts.RADAR, arrDataStudents, optsChartRadar10);
 
 		// BTM-R: Filled/Axis Options
-		slide.addText(".", { x: 6.83, y: 3.8, w: 6.0, h: 3.5, fill: { color: "F1F1F1" }, color: "F1F1F1" });
 		let optsChartRadar11 = {
 			x: 6.83,
 			y: 3.8,
 			w: 6.0,
 			h: 3.5,
+			chartArea: { fill: { color: "F1F1F1" } },
+			//
 			radarStyle: "filled",
 			//
 			chartColors: [COLOR_RED, COLOR_AMB, COLOR_GRN], // marker & line color
@@ -1564,13 +1648,14 @@ function genSlide17(pptx) {
 	slide.addNotes("API Docs: https://gitbrent.github.io/PptxGenJS/docs/api-charts.html");
 	slide.addTable([[{ text: "Chart Examples: Multi-Type Charts", options: BASE_TEXT_OPTS_L }, BASE_TEXT_OPTS_R]], BASE_TABLE_OPTS);
 
+	// TOP-R:
 	function doStackedLine() {
-		// TOP-RIGHT:
 		let opts = {
-			x: 7.0,
+			x: 6.83,
 			y: 0.6,
 			w: 6.0,
 			h: 3.0,
+			chartArea: { fill: { color: "F1F1F1" } },
 			barDir: "col",
 			barGrouping: "stacked",
 			catAxisLabelColor: "0000CC",
@@ -1625,12 +1710,15 @@ function genSlide17(pptx) {
 		slide.addChart(chartTypes, opts);
 	}
 
+	// TOP-L:
 	function doColumnAreaLine() {
 		let opts = {
-			x: 0.6,
+			x: 0.5,
 			y: 0.6,
 			w: 6.0,
 			h: 3.0,
+			chartArea: { fill: { color: "F1F1F1" } },
+
 			barDir: "col",
 			catAxisLabelColor: "666666",
 			catAxisLabelFontFace: "Arial",
@@ -1714,13 +1802,15 @@ function genSlide17(pptx) {
 		slide.addChart(chartTypes, opts);
 	}
 
+	// BTM-L:
 	function doStackedDot() {
-		// BOT-LEFT:
 		let opts = {
-			x: 0.6,
+			x: 0.5,
 			y: 4.0,
 			w: 6.0,
 			h: 3.0,
+			chartArea: { fill: { color: "F1F1F1" } },
+
 			barDir: "col",
 			barGrouping: "stacked",
 			catAxisLabelColor: "999999",
@@ -1810,13 +1900,15 @@ function genSlide17(pptx) {
 		slide.addChart(chartTypes, opts);
 	}
 
+	// BTM-R:
 	function doBarCol() {
-		// BOT-RGT:
 		let opts = {
-			x: 7,
+			x: 6.83,
 			y: 4.0,
 			w: 6.0,
 			h: 3.0,
+			chartArea: { fill: { color: "F1F1F1" } },
+
 			barDir: "col",
 			barGrouping: "stacked",
 			catAxisLabelColor: "999999",
@@ -1890,6 +1982,7 @@ function genSlide17(pptx) {
 				options: {
 					barDir: "bar",
 					barGrouping: "standard",
+					chartColors: ["0077BF", "4E9D2D", "ECAA00", "5FC4E3", "DE4216", "154384"],
 					secondaryValAxis: !!opts.valAxes,
 					secondaryCatAxis: !!opts.catAxes,
 				},
@@ -1965,10 +2058,10 @@ function genSlide17(pptx) {
 		slide.addChart(chartTypes, multiOpts);
 	}
 
-	doBarCol();
-	doStackedDot();
 	doColumnAreaLine();
 	doStackedLine();
+	doStackedDot();
+	doBarCol();
 	//readmeExample();
 }